+++ src/main/kotlin/crossref/ManyToManyRelation.kt
@@ -4,6 +4,6 @@ interface ManyToManyRelation<Left : Any, Right : Any> {
     fun add(left: Left, right: Right)
     fun remove(left: Left, right: Right)
 
-    fun getRights(left: Left): Collection<Right>
-    fun getLefts(right: Right): Collection<Left>
+    fun getRight(left: Left): Set<Right>
+    fun getLeft(right: Right): Set<Left>
 }