+++ src/main/kotlin/crossref/OneToManyRelation.kt
@@ -4,5 +4,5 @@ interface OneToManyRelation<Left : Any, Right : Any> {
     operator fun set(right: Right, left: Left?)
 
     fun getLeft(right: Right): Left?
-    fun getRights(left: Left): Collection<Right>
+    fun getRight(left: Left): Set<Right>
 }