+++ src/main/kotlin/crossref/impl/SimpleOneToOneRelation.kt
@@ -4,40 +4,39 @@ import crossref.OneToOneRelation
 
 open class SimpleOneToOneRelation<Left : Any, Right : Any> : OneToOneRelation<Left, Right> {
 
-    private val rightByLeft = HashMap<Int, Right>()
-    private val leftByRight = HashMap<Int, Left>()
+    private val rightByLeft = HashMap<Left, Right>()
+    private val leftByRight = HashMap<Right, Left>()
 
-    override final fun set(left: Left?, right: Right?): Unit = synchronized(this) {
+    @Synchronized
+    override final fun set(left: Left?, right: Right?) {
         if (left != null) {
-            val id = idOfLeft(left)
-
-            rightByLeft[id]?.let {
-                leftByRight -= idOfRight(it)
+            rightByLeft[left]?.let {
+                leftByRight -= it
             }
 
             if (right != null)
-                rightByLeft[id] = right
+                rightByLeft[left] = right
             else
-                rightByLeft -= id
+                rightByLeft -= left
         }
 
         if (right != null) {
-            val id = idOfRight(right)
-
-            leftByRight[id]?.let {
-                rightByLeft -= idOfLeft(it)
+            leftByRight[right]?.let {
+                rightByLeft -= it
             }
 
             if (left != null)
-                leftByRight[id] = left
+                leftByRight[right] = left
             else
-                leftByRight -= id
+                leftByRight -= right
         }
     }
 
-    override final fun getLeft(right: Right): Left? = leftByRight[idOfRight(right)]
-    override final fun getRight(left: Left): Right? = rightByLeft[idOfLeft(left)]
+    override final fun getLeft(right: Right): Left? = synchronized(this) {
+        leftByRight[right]
+    }
 
-    open fun idOfLeft(left: Left): Int = System.identityHashCode(left)
-    open fun idOfRight(right: Right): Int = System.identityHashCode(right)
+    override final fun getRight(left: Left): Right? = synchronized(this) {
+        rightByLeft[left]
+    }
 }