+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 ## 1.0-SNAPSHOT (unreleased)
 ### Added
 * Base relation interfaces `OneToOneRelation`, `OneToManyRelation` and `ManyToManyRelation`
-* Base implementation of `OneToOneRelation`
-* Properties for `OneToOneRelation`
+* Base implementation of `OneToOneRelation` and `OneToManyRelation`
+* Properties for `OneToOneRelation` and `OneToManyRelation`